requirements.txt
flask
gunicorn
pandas
openai
python-dotenv
numpy